Standardise code action test
