+++ compiler-plugin-tests/src/test/resources/testng.xml
@@ -23,7 +23,7 @@
     <test name="CDC Compiler Plugin Tests">
         <classes>
             <class name="io.ballerina.lib.cdc.compiler.CompilerPluginTest"/>
-            <class name="io.ballerina.lib.cdc.compiler.CodeSnippetGenerationCodeActionTest"></class>
+            <class name="io.ballerina.lib.cdc.compiler.CodeActionTest"/>
         </classes>
     </test>
 </suite>
